+++ src/main/java/com/pharm/repository/dci/DCIRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DCIRepository extends JpaRepository<DCI,Integer> {
+public interface DCIRepository extends JpaRepository<DCI,Long> {
 }
