+++ src/main/java/com/pharm/service/interfaces/type/TypeService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface TypeService {
     List<Type> findAll();
-    Type findById(Integer id);
+    Type findById(Long id);
     Type create(Type type);
     Type update(Type type);
     Type delete(Type type);
